@@ -1,6 +1,6 @@
 {
     "name": "electroncloudmusic",
-    "version": "0.12.0",
+    "version": "0.13.0",
     "private": true,
     "scripts": {
         "serve": "vue-cli-service serve",
@@ -14,34 +14,36 @@
     "main": "background.js",
     "dependencies": {
         "axios": "^0.21.1",
-        "core-js": "^3.6.5",
-        "element-plus": "^1.0.2-beta.44",
-        "vue": "^3.0.0",
+        "core-js": "^3.15.2",
+        "element-plus": "^1.0.2-beta.54",
+        "vue": "^3.1.4",
         "vue-router": "^4.0.0-0",
         "vuex": "^4.0.0-0"
     },
     "devDependencies": {
         "@types/electron-devtools-installer": "^2.2.0",
-        "@vue/cli-plugin-babel": "~4.5.0",
-        "@vue/cli-plugin-router": "~4.5.0",
-        "@vue/cli-plugin-typescript": "~4.5.0",
-        "@vue/cli-plugin-vuex": "~4.5.0",
-        "@vue/cli-service": "^4.5.0",
-        "@vue/compiler-sfc": "^3.0.0",
-        "electron-devtools-installer": "^3.1.0",
+        "@vue/cli-plugin-babel": "~4.5.13",
+        "@vue/cli-plugin-router": "~4.5.13",
+        "@vue/cli-plugin-typescript": "~4.5.13",
+        "@vue/cli-plugin-vuex": "~4.5.13",
+        "@vue/cli-service": "^4.5.13",
+        "@vue/compiler-sfc": "^3.1.4",
+        "electron": "^13.1.5",
+        "electron-devtools-installer": "^3.2.0",
         "jsonfile": "^6.1.0",
-        "sass": "^1.26.5",
-        "sass-loader": "^8.0.2",
-        "typescript": "~4.1.5",
-        "vite": "^2.2.4",
-        "vite-plugin-style-import": "^0.10.1",
+        "sass": "^1.35.1",
+        "sass-loader": "^12.1.0",
+        "typescript": "~4.3.5",
+        "vite": "^2.3.3",
+        "vite-plugin-style-import": "^1.0.1",
         "vue-class-component": "^8.0.0-0",
-        "vue-cli-plugin-electron-builder": "*",
+        "vue-cli-plugin-electron-builder": "^2.1.1",
         "vue-cli-plugin-vite": "~1.1.4"
     },
     "browserslist": [
         "> 1%",
         "last 2 versions",
         "not dead"
-    ]
+    ],
+    "__npminstall_done": false
 }
