@@ -1,15 +1,16 @@
-'use strict'
+import {
+  app, protocol, BrowserWindow, ipcMain,
+} from 'electron';
+import { createProtocol } from 'vue-cli-plugin-electron-builder/lib';
+import installExtension, { VUEJS_DEVTOOLS } from 'electron-devtools-installer';
 
-import { app, protocol, BrowserWindow,ipcMain } from 'electron'
-import { createProtocol } from 'vue-cli-plugin-electron-builder/lib'
-import installExtension, { VUEJS_DEVTOOLS } from 'electron-devtools-installer'
-const isDevelopment = process.env.NODE_ENV !== 'production'
-const fs = require("fs");
+const isDevelopment = process.env.NODE_ENV !== 'production';
+const fs = require('fs');
 
 // Scheme must be registered before the app is ready
 protocol.registerSchemesAsPrivileged([
-  { scheme: 'app', privileges: { secure: true, standard: true } }
-])
+  { scheme: 'app', privileges: { secure: true, standard: true } },
+]);
 
 async function createWindow() {
   // Create the browser window.
@@ -17,24 +18,24 @@ async function createWindow() {
     width: 1280,
     height: 768,
     webPreferences: {
-      nodeIntegration:true,
-      contextIsolation:false,
-      preload: 'app://./preload.js'
+      nodeIntegration: true,
+      contextIsolation: false,
+      preload: 'app://./preload.js',
       // Use pluginOptions.nodeIntegration, leave this alone
       // See nklayman.github.io/vue-cli-plugin-electron-builder/guide/security.html#node-integration for more info
       // nodeIntegration: (process.env
       //     .ELECTRON_NODE_INTEGRATION as unknown) as boolean
-    }
-  })
+    },
+  });
 
   if (process.env.WEBPACK_DEV_SERVER_URL) {
     // Load the url of the dev server if in development mode
-    await win.loadURL(process.env.WEBPACK_DEV_SERVER_URL as string)
-    if (!process.env.IS_TEST) win.webContents.openDevTools()
+    await win.loadURL(process.env.WEBPACK_DEV_SERVER_URL as string);
+    if (!process.env.IS_TEST) win.webContents.openDevTools();
   } else {
-    createProtocol('app')
+    createProtocol('app');
     // Load the index.html when not in development
-    win.loadURL('app://./index.html')
+    win.loadURL('app://./index.html');
   }
 }
 
@@ -43,15 +44,15 @@ app.on('window-all-closed', () => {
   // On macOS it is common for applications and their menu bar
   // to stay active until the user quits explicitly with Cmd + Q
   if (process.platform !== 'darwin') {
-    app.quit()
+    app.quit();
   }
-})
+});
 
 app.on('activate', () => {
   // On macOS it's common to re-create a window in the app when the
   // dock icon is clicked and there are no other windows open.
-  if (BrowserWindow.getAllWindows().length === 0) createWindow()
-})
+  if (BrowserWindow.getAllWindows().length === 0) createWindow();
+});
 
 // This method will be called when Electron has finished
 // initialization and is ready to create browser windows.
@@ -60,38 +61,38 @@ app.on('ready', async () => {
   if (isDevelopment && !process.env.IS_TEST) {
     // Install Vue Devtools
     try {
-      await installExtension(VUEJS_DEVTOOLS)
+      await installExtension(VUEJS_DEVTOOLS);
     } catch (e) {
-      console.error('Vue Devtools failed to install:', e.toString())
+      console.error('Vue Devtools failed to install:', e.toString());
     }
   }
-  createWindow()
-})
+  createWindow();
+});
 
 // Exit cleanly on request from parent process in development mode.
 if (isDevelopment) {
   if (process.platform === 'win32') {
     process.on('message', (data) => {
       if (data === 'graceful-exit') {
-        app.quit()
+        app.quit();
       }
-    })
+    });
   } else {
     process.on('SIGTERM', () => {
-      app.quit()
-    })
+      app.quit();
+    });
   }
 }
 
-//receiver settingdata from render process
-ipcMain.on("setting-save",(event,args)=>{
+// receiver settingdata from render process
+ipcMain.on('setting-save', (event, args) => {
   console.log(args);
-  fs.open("C:\\Users\\empty\\Desktop\\setting","w",(err:Error,fd:number)=>{
-    if(err){
-      console.log(err);      
-    }else{
-      fs.writeFile(fd,args);
+  fs.open('C:\\Users\\empty\\Desktop\\setting', 'w', (err:Error, fd:number) => {
+    if (err) {
+      console.log(err);
+    } else {
+      fs.writeFile(fd, args);
       fs.close(fd);
     }
-  })
-})
+  });
+});
