@@ -1,5 +1,5 @@
-import { ComponentCustomProperties } from 'vue'
-import { Store } from "@/store";
+import { ComponentCustomProperties } from 'vue';
+import { Store } from '@/store';
 
 declare module '@vue/runtime-core' {
   interface State{
