@@ -1,8 +1,9 @@
-import { createApp } from 'vue'
-import App from './App.vue'
-import router from './router'
-import store from './store'
-import ElementPlus from 'element-plus'
-import 'element-plus/lib/theme-chalk/index.css'
+import { createApp } from 'vue';
+import ElementPlus from 'element-plus';
+import App from './App.vue';
+import router from './router';
+import store from './store';
+import 'element-plus/lib/theme-chalk/index.css';
 
-createApp(App).use(store).use(router).use(ElementPlus).mount('#app');
+createApp(App).use(store).use(router).use(ElementPlus)
+  .mount('#app');
